refactor(validate): split SLI validation code
